talentprobe_be/app/__init__.py
from pathlib import Path

from dotenv import load_dotenv


# Load environment variables for all backend entrypoints (uvicorn, scripts, tests).
load_dotenv(Path(__file__).resolve().parents[1] / ".env")
