talentprobe_be/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
python-multipart==0.0.20
pypdf==5.2.0
python-docx==1.1.2
mysql-connector-python==9.1.0
python-dotenv==1.0.1
google-auth==2.38.0
requests==2.32.3
boto3==1.37.38
